Refactor ns_string
