+++ crates/icrate/src/Foundation/__macro_helpers/mod.rs
@@ -1,0 +1,3 @@
+mod ns_string;
+
+pub use self::ns_string::*;
