+++ crates/icrate/src/Foundation/macros/mod.rs
@@ -1,0 +1,1 @@
+mod ns_string;
